+++ src/components/Internships/AllInternships.js
@@ -78,33 +78,7 @@ const AllInternships = () => {
       });
   };
 
-  const bookMarkPost = (postId) => {
-    axios({
-      method: "post",
-      url: `http://localhost:5000/student/bookmarkInternship/${postId}`,
-      headers: {
-        Authorization: "Bearer " + localStorage.getItem("jwt"),
-        "Content-Type": "application/json",
-      },
-    })
-      .then((res) => {
-        console.log(res);
-        if (res.data.error) {
-          // console.log(res.data.error);
-          const notify = () => toast(res.data.error);
-          notify();
-        } else {
-          // setInternships(res.data.internships);
-          // window.location.reload(false);
-          console.log(res.data.message);
-          const notify = () => toast(res.data.message);
-          notify();
-        }
-      })
-      .catch((err) => {
-        console.log("Error: ", err);
-      });
-  };
+
 
   return (
     <div className="internshipsOuterContainer">
@@ -146,7 +120,6 @@ const AllInternships = () => {
                   internship={internship}
                   userId={state.user._id}
                   deletePost={deletePost}
-                  bookMarkPost={bookMarkPost}
                 />
               </Col>
             );
